refactor(js_syntax): share the function-boundary predicate
Moved from `biome_js_analyze::ast_utils` unchanged, so `biome_js_type_info`
can use it without depending on the analyzer. The kinds that can never be
`async` get a name of their own, since two callers want only those.
